remove some compile warnings

+++ src/git.cpp
@@ -219,10 +219,10 @@ QModelIndex FileHistory::parent(const QModelIndex&) const {
 
 const QString FileHistory::timeDiff(unsigned long secs) const {
 
-	uint days  =  secs / (3600 * 24);
-	uint hours = (secs - days * 3600 * 24) / 3600;
-	uint min   = (secs - days * 3600 * 24 - hours * 3600) / 60;
-	uint sec   =  secs - days * 3600 * 24 - hours * 3600 - min * 60;
+        ulong days  =  secs / (3600 * 24);
+        ulong hours = (secs - days * 3600 * 24) / 3600;
+        ulong min   = (secs - days * 3600 * 24 - hours * 3600) / 60;
+        ulong sec   =  secs - days * 3600 * 24 - hours * 3600 - min * 60;
 	QString tmp;
 	if (days > 0)
 		tmp.append(QString::number(days) + "d ");
